refactor: reorganize types into semantic files; add AGENTS.md entrypoint
