--- src/types.ts
@@ -1,217 +1,0 @@
-import {
-  FileChild, IParagraphStylePropertiesOptions, IPropertiesOptions, IRunStylePropertiesOptions,
-  IShadingAttributesProperties, ParagraphChild
-} from 'docx'
-import { MarkedOptions, Tokens } from 'marked'
-
-import { BlockKatex, Footnote, FootnoteRef, InlineKatex } from './extensions'
-import { MarkdownDocx } from './MarkdownDocx'
-
-export type MarkdownImageType = 'jpg' | 'png' | 'gif' | 'bmp'
-
-export type MarkdownImageItem = {
-  type: MarkdownImageType
-  data: Buffer | string | Uint8Array | ArrayBuffer
-  width: number
-  height: number
-}
-
-export type MarkdownImageAdapter = (token: Tokens.Image) => Promise<null | MarkdownImageItem>
-
-export interface MarkdownDocxOptions extends MarkedOptions {
-  imageAdapter?: MarkdownImageAdapter
-
-  /**
-   * Math engine configuration
-   * builtin: simple unicode mapping
-   * katex: KaTeX -> MathML -> docx Math
-   */
-  math?: {
-    engine?: 'builtin' | 'katex' // future: 'mathjax'
-    katexOptions?: Record<string, any>
-    /** Prefer constructs that are broadly supported by LibreOffice (e.g., avoid true OMML matrices and n-ary) */
-    libreOfficeCompat?: boolean
-  }
-
-  /**
-   * do not download image
-   * @default false
-   */
-  ignoreImage?: boolean
-
-  /**
-   * do not parse footnote
-   * @default false
-   */
-  ignoreFootnote?: boolean
-
-  /**
-   * do not parse html
-   * @default false
-   */
-  ignoreHtml?: boolean
-
-  /**
-   * Properties for the document
-   */
-  document?: Omit<IPropertiesOptions, 'sections'>
-
-  /**
-   * colors theme
-   */
-  theme?: Partial<IMarkdownTheme>
-}
-
-export type IBlockToken =
-  | Tokens.Space
-  | Tokens.Code
-  | Tokens.Heading
-  | Tokens.Hr
-  | Tokens.Blockquote
-  | Tokens.List
-  | Tokens.HTML
-  | Tokens.Def
-  | Tokens.Table
-  | Tokens.Heading
-  | Tokens.Paragraph
-  | Tokens.Text
-  // plugin
-  | Footnote
-
-export type IInlineToken =
-  | Tokens.Escape
-  | Tokens.Tag
-  | Tokens.Link
-  | Tokens.Em
-  | Tokens.Strong
-  | Tokens.Codespan
-  | Tokens.Br
-  | Tokens.Del
-  | Tokens.Text
-  | Tokens.Image
-  // plugin
-  | FootnoteRef
-  | InlineKatex
-  | BlockKatex
-
-export type IParagraphToken =
-  | Tokens.Paragraph
-  | Tokens.Blockquote
-  | Tokens.Heading
-
-export type ITextAttr = {
-  style?: string
-
-  // attrs
-  bold?: boolean
-  italics?: boolean
-  underline?: boolean // with options
-  strike?: boolean
-  break?: boolean | number
-
-  // text style
-  html?: boolean
-  link?: boolean
-  strong?: boolean
-  em?: boolean
-  codespan?: boolean
-  del?: boolean
-  br?: boolean
-
-}
-
-export type IBlockAttr = {
-  style?: string
-
-  blockquote?: boolean
-
-  list?: {
-    task?: boolean
-    checked?: boolean
-    level: number
-    type?: 'number' | 'bullet'
-    /**
-     * @link https://github.com/dolanmiu/docx/pull/816
-     * @link https://github.com/dolanmiu/docx/issues/3037#issuecomment-3164253396
-     */
-    instance?: number // numbering instance
-  }
-
-  listNone?: boolean
-
-  heading?: number
-  code?: boolean
-
-  align?: 'left' | 'center' | 'right' | null
-
-  footnote?: boolean
-
-}
-
-export type Writeable<T> = {
-  -readonly [P in keyof T]: T[P]
-}
-
-export type IMarkdownToken =
-  | 'space' | 'code' | 'hr' | 'blockquote' | 'html' | 'def' | 'paragraph' | 'text' | 'footnote' | 'listItem' | 'table' | 'tableHeader' | 'tableCell' | 'heading1' | 'heading2' | 'heading3' | 'heading4' | 'heading5' | 'heading6'
-  | 'tag' | 'link' | 'strong' | 'em' | 'codespan' | 'del' | 'br'
-
-export type IMarkdownStyle = {
-  inline?: boolean
-  className: string
-  name?: string
-  basedOn?: string
-  next?: string
-  run?: IRunStylePropertiesOptions
-  paragraph?: IParagraphStylePropertiesOptions
-  quickFormat?: boolean
-
-  // special attributes
-  properties?: any
-}
-
-export type IMarkdownRenderFunction =
-  (render: MarkdownDocx, token: IInlineToken | IBlockToken, attr?: ITextAttr | IBlockAttr)
-    => ParagraphChild | ParagraphChild[] | FileChild | FileChild[] | false | null
-
-// color themes
-export type IMarkdownTheme = {
-  heading1: string
-  heading2: string
-  heading3: string
-  heading4: string
-  heading5: string
-  heading6: string
-  link: string
-  code: string
-  tag: string
-  border: string
-  codespan: string
-  codeBackground: string
-  blockquote: string
-  blockquoteBackground: string
-  del: string
-  hr: string
-  html: string
-  tableHeaderBackground: string
-
-  /**
-   * size of heading fonts
-   */
-  heading1Size: number
-  heading2Size: number
-  heading3Size: number
-  heading4Size: number
-  heading5Size: number
-  heading6Size: number
-  spaceSize: number
-  codeSize: number
-  linkUnderline: boolean
-
-  /**
-   * Body typography customization
-   */
-  bodySize?: number        // Font size in points (e.g., 14 for 14pt). Optional, uses library default if omitted.
-  lineSpacing?: number     // Line spacing multiplier (e.g., 1.5 for 150%). Optional, uses library default if omitted.
-
-}
