+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export type { IMarkdownTheme } from './theme'
+export type { MarkdownImageType, MarkdownImageItem, MarkdownImageAdapter } from './image'
+export type { IBlockToken, IInlineToken, IParagraphToken } from './token'
+export type { ITextAttr, IBlockAttr } from './attr'
+export type { Writeable, IMarkdownToken, IMarkdownStyle, IMarkdownRenderFunction } from './style'
+export type { MarkdownDocxOptions } from './options'
